git-annex.rb: Do not auto-start git-annex assistant
This commit prevents the git-annex formula from autostarting the git-annex assistant.

Git-annex assistant is a daemon that automatically creates commits in git-annex repositories *and* automatically syncs them with remotes. See <https://git-annex.branchable.com/assistant/>.

In my experience, the git-annex assistant is usually an opt-in feature. I was surprised when I cloned a git-annex repo to my Macbook and found commits were being created automatically and synced to my server.